First attempt to permanently dismiss configuration workout notice
