+++ inc/options/class-wpseo-option-wpseo.php
@@ -81,6 +81,7 @@ class WPSEO_Option_Wpseo extends WPSEO_Option {
 		'algolia_integration_active'               => false,
 		'import_cursors'                           => [],
 		'workouts_data'                            => [ 'configuration' => [ 'finishedSteps' => [] ] ],
+		'dismiss_configuration_workout_notice'     => false,
 	];
 
 	/**
